Add clean-docs and apidocs
I want to be able to change the apidocs so I don't want to them to
be regenerated each time. In particular I'd like to change how a
"sources" are displayed. Currently the window is not wide enough.

Also, this just feels better, could be wrong, we'll see over time
how it works out.