Cleanup documentation
Add sphinx-prompt and sphinx_substitution_extensions to conf.py. And also
add rst_prolog with the replacement string.

Use prompt:: instead of code-block:: where we have command line scripts
that should be copied and pasted. Doing this has the benefit of not
copying the `$`.

Add pagedocs.rst with version

readme.rst is no longer included in base page of docs

HiSTORY.rst isn't being updated so just remove for now.